refactor: split rates decomposition view
